requirements.txt
Django
django-stubs[compatible-mypy]
django-stubs-ext
httpx
python-dotenv
djangorestframework
djangorestframework-stubs[compatible-mypy]